requirements.txt
django==1.10
djangorestframework
Pillow
